requirements.txt
numpy==1.16.2
mock==3.0.5
opencv-python==4.1.2.30
opencv-contrib-python==4.1.2.30
scipy==1.2.2
matplotlib==2.2.4
pillow==6.2.1
easydict==1.9
h5py==2.10.0
PyYAML==5.1.1
cython==0.24
keras==2.1.5
